12
new tag converter
about to go for gui

+++ src/domainLogic/util/Bitrate.java
@@ -6,8 +6,14 @@ public class Bitrate {
 
     private BigDecimal bitrate;
 
-    public Bitrate(BigDecimal bitrate) { this.bitrate = bitrate; }
-    public Bitrate(){ this.bitrate = BigDecimal.valueOf(0.0); }
+    public Bitrate(double bitrate) {
+
+        this.bitrate = BigDecimal.valueOf(bitrate);
+    }
+
+    public Bitrate(){
+        this.bitrate = BigDecimal.valueOf(0.0);
+    }
 
     public BigDecimal getBitrate() {
         return bitrate;
